added: [ 1781132 ] Various patches to make it compile on 64 bit archs - unrar.patch. thanks to seppel
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk/XBMC@10084 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/lib/UnrarXLib/crc.cpp
@@ -20,7 +20,7 @@ uint CRC(uint StartCRC,const void *Addr,uint Size)
     InitCRC();
   byte *Data=(byte *)Addr;
 #if defined(LITTLE_ENDIAN) && defined(PRESENT_INT32)
-  while (Size>0 && ((int)Data & 7))
+  while (Size>0 && ((intptr_t)Data & 7))
   {
     StartCRC=CRCTab[(byte)(StartCRC^Data[0])]^(StartCRC>>8);
     Size--;
